backend/requirements.txt
pyproject-toml
psycopg2
aiohttp==3.8.2
yarl==1.8.1
frozenlist==1.3.1
Django
environs
gunicorn
pandas
djangorestframework
openpyxl
requests
django-admin-rangefilter
django-filter
